Committed changes for Homework 10
Added the Major class and associated processing. Students are updated to track the remaining required and elective courses for their major. The Student class also knows when the student has satisfied the elective requirement. The tables output by the tool now include all majors and the remaining required and elective courses for each student.